GraspAndMotionet/requirements.txt
scikit-image 
matplotlib 
imageio 
plotly 
opencv-python
black 
usort 
flake8 
flake8-bugbear 
flake8-comprehensions
open3d==0.13.0
trimesh
omegaconf
hydra-core
open3d==0.13.0
pytorch-kinematics
urdf_parser_py
loguru
colorama
tensorboard
transforms3d
pytransform3d
git+'https://github.com/otaheri/chamfer_distance'
git+'https://github.com/otaheri/MANO' 